+++ src/SshConnectDialog.h
@@ -6,6 +6,8 @@ class QLineEdit;
 class QSpinBox;
 class QCheckBox;
 class QRadioButton;
+class QListWidget;
+class QPushButton;
 
 class SshConnectDialog : public QDialog
 {
@@ -14,7 +16,18 @@ public:
     explicit SshConnectDialog(QWidget *parent = nullptr);
     SshHostInfo hostInfo() const;
 
+private slots:
+    void saveConnection();
+    void deleteConnection();
+    void loadSelectedConnection();
+
 private:
+    void loadSavedConnections();
+    void populateList();
+
+    QListWidget *m_connectionList;
+    QPushButton *m_deleteBtn;
+
     QLineEdit *m_host;
     QSpinBox *m_port;
     QLineEdit *m_username;
@@ -22,4 +35,14 @@ private:
     QLineEdit *m_keyPath;
     QRadioButton *m_passwordAuth;
     QRadioButton *m_keyAuth;
+
+    struct SavedConnection {
+        QString name;
+        QString host;
+        int port = 22;
+        QString username;
+        QString privateKeyPath;
+        bool useKeyAuth = false;
+    };
+    QList<SavedConnection> m_saved;
 };